Move OrganizationPlanType to Billing.Organizations.Models
