+++ src/Core/AdminConsole/Repositories/IOrganizationRepository.cs
@@ -1,7 +1,7 @@
 ﻿using System.Data.Common;
 using Bit.Core.AdminConsole.Entities;
 using Bit.Core.AdminConsole.Enums.Provider;
-using Bit.Core.AdminConsole.Models.Data;
+using Bit.Core.Billing.Organizations.Models;
 using Bit.Core.Models.Data.Organizations;
 using Bit.Core.Models.Data.Organizations.OrganizationUsers;
