+++ src/Core/Billing/Organizations/Models/OrganizationPlanType.cs
@@ -1,6 +1,6 @@
 ﻿using Bit.Core.Billing.Enums;
 
-namespace Bit.Core.AdminConsole.Models.Data;
+namespace Bit.Core.Billing.Organizations.Models;
 
 /// <summary>Lightweight projection of <c>(Organization.Id, Organization.PlanType)</c>.</summary>
 public class OrganizationPlanType
