+++ src/Infrastructure.Dapper/AdminConsole/Repositories/OrganizationRepository.cs
@@ -2,8 +2,8 @@
 using System.Data.Common;
 using Bit.Core.AdminConsole.Entities;
 using Bit.Core.AdminConsole.Enums.Provider;
-using Bit.Core.AdminConsole.Models.Data;
 using Bit.Core.Auth.Entities;
+using Bit.Core.Billing.Organizations.Models;
 using Bit.Core.Entities;
 using Bit.Core.Models.Data.Organizations;
 using Bit.Core.Models.Data.Organizations.OrganizationUsers;
