+++ test/Core.Test/Billing/Organizations/PlanMigration/Commands/BulkSyncCohortAssignmentsCommandTests.cs
@@ -1,5 +1,5 @@
-﻿using Bit.Core.AdminConsole.Models.Data;
-using Bit.Core.Billing.Enums;
+﻿using Bit.Core.Billing.Enums;
+using Bit.Core.Billing.Organizations.Models;
 using Bit.Core.Billing.Organizations.PlanMigration.Commands;
 using Bit.Core.Billing.Organizations.PlanMigration.Entities;
 using Bit.Core.Billing.Organizations.PlanMigration.Enums;
